+++ CHANGELOG.md
@@ -1,3 +1,35 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial template
+
+## 1.0.0
+
+* Added flavor banner, and a utils class to recognize flavor build type
+
+## 1.0.1
+
+* Fix on missing dependency
+
+## 1.0.2
+
+* Removed unuseful code
+
+## 1.1.0
+
+* Flavor banner parameters improvement
+
+## 2.0.0
+
+* Updated flutter to 3.0.2
+
+## 2.0.1
+
+* Minor fixes
+
+## 2.0.2
+
+* Minor fixes
+
+## 2.0.3
+
+* Minor fix on dialog opening